doGet test for loginServlet; removed unneeded imports in SearchServletTest
